@@ -22,24 +22,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-mdc/web": "^3.2.1",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular-mdc/web": "^5.1.1",
+    "@angular/animations": "^9.1.12",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/flex-layout": "^9.0.0-beta.31",
+    "@angular/forms": "^9.1.12",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "^9.1.12",
+    "@angular/platform-browser-dynamic": "^9.1.12",
     "@angular/pwa": "^0.803.26",
-    "@angular/router": "^8.2.14",
-    "@angular/service-worker": "^8.2.14",
-    "@bugsnag/js": "^6.4.0",
-    "@ngneat/transloco": "^2.5.2",
+    "@angular/router": "^9.1.12",
+    "@angular/service-worker": "^9.1.12",
+    "@bugsnag/js": "^7.3.5",
+    "@ngneat/transloco": "^2.19.1",
     "@ngneat/transloco-locale": "^1.0.0",
     "@sillsdev/machine": "^2.1.5",
-    "angular-file": "^1.3.2",
+    "angular-file": "^3.1.2",
     "angular-split": "^3.0.2",
     "arraydiff": "^0.1.3",
     "auth0-js": "^9.10.4",
@@ -59,21 +59,22 @@
     "process": "^0.11.10",
     "quill": "^1.3.7",
     "realtime-server": "file:../../RealtimeServer",
-    "reconnecting-websocket": "^4.2.0",
+    "reconnecting-websocket": "^4.4.0",
     "recordrtc": "^5.5.7",
     "rich-text": "^3.1.0",
-    "rxjs": "6.5.2",
+    "rxjs": "6.6.2",
     "sharedb": "^1.0.0-beta.22",
+    "tslib": "^1.13.0",
     "uuid": "^3.3.2",
     "xregexp": "^4.2.4",
-    "zone.js": "^0.9.1"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular/cli": "^8.3.26",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "^0.901.12",
+    "@angular/cli": "^9.1.12",
+    "@angular/compiler": "^9.1.12",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/language-service": "^9.1.12",
     "@ngneat/transloco-keys-manager": "^1.0.0",
     "@types/auth0-js": "^9.10.7",
     "@types/jasmine": "~3.3.12",
@@ -96,14 +97,13 @@
     "karma-jasmine-html-reporter": "^1.4.2",
     "karma-teamcity-reporter": "^1.1.0",
     "lint-staged": "^10.2.11",
-    "ng-bullet": "^1.0.3",
     "npm": "6.14.6",
-    "prettier": "1.15.3",
+    "prettier": "^2.1.1",
     "protractor": "^5.4.3",
     "ts-mockito": "^2.5.0",
     "ts-node": "~8.8.1",
     "tslint": "^5.20.1",
-    "typescript": "~3.5.3",
+    "typescript": "~3.8.3",
     "webpack-bundle-analyzer": "^3.6.1"
   },
   "husky": {
