@@ -15,14 +15,11 @@ exports.config = {
   jasmineNodeOpts: {
     showColors: true,
     defaultTimeoutInterval: 30000,
-    print: function() {}
+    print: function () {}
   },
   SELENIUM_PROMISE_MANAGER: false,
   onPrepare() {
-    browser.driver
-      .manage()
-      .window()
-      .maximize();
+    browser.driver.manage().window().maximize();
 
     require('ts-node').register({
       project: 'e2e/tsconfig.e2e.json'
