@@ -560,7 +560,10 @@ class TestEnvironment {
           MAT: {
             bookNum: 40,
             hasSource: false,
-            chapters: [{ number: 1, lastVerse: 25, isValid: true }, { number: 2, lastVerse: 23, isValid: true }]
+            chapters: [
+              { number: 1, lastVerse: 25, isValid: true },
+              { number: 2, lastVerse: 23, isValid: true }
+            ]
           },
           LUK: { bookNum: 42, hasSource: false, chapters: [{ number: 1, lastVerse: 80, isValid: true }] },
           JHN: { bookNum: 43, hasSource: false, chapters: [{ number: 1, lastVerse: 0, isValid: true }] }
