@@ -160,7 +160,10 @@ class TestEnvironment {
   matthewText: TextInfo = {
     bookNum: 40,
     hasSource: false,
-    chapters: [{ number: 1, lastVerse: 25, isValid: true }, { number: 3, lastVerse: 17, isValid: true }]
+    chapters: [
+      { number: 1, lastVerse: 25, isValid: true },
+      { number: 3, lastVerse: 17, isValid: true }
+    ]
   };
   readonly PROJECT01: string = 'project01';
   adminUser: UserInfo = { id: 'user01', role: SFProjectRole.ParatextAdministrator };
