@@ -276,7 +276,10 @@ class TestEnvironment {
         sync: { queuedCount: 0 },
         texts:
           args.hasTexts == null || args.hasTexts
-            ? [{ bookNum: 40, chapters: [], hasSource: false }, { bookNum: 41, chapters: [], hasSource: false }]
+            ? [
+                { bookNum: 40, chapters: [], hasSource: false },
+                { bookNum: 41, chapters: [], hasSource: false }
+              ]
             : [],
         userRoles: { user01: args.role == null ? SFProjectRole.ParatextTranslator : args.role }
       }
