@@ -152,9 +152,7 @@ class TestEnvironment {
 class DialogTestModule {}
 
 @Component({
-  template: `
-    <button (click)="openDialog()"></button>
-  `
+  template: `<button (click)="openDialog()"></button>`
 })
 class DialogOpenerComponent {
   publicName: string = 'Simon Says';
