@@ -1,38 +1,52 @@
+import { Event } from '@bugsnag/js';
 import { ErrorReportingService } from './error-reporting.service';
 
-interface Breadcrumb {
-  type: string;
-  metaData: {
-    from: string;
-    to: string;
-  };
-}
-
 describe('ErrorReportingService', () => {
-  it('should redact the access_token from the breadcrumb URL', async () => {
-    const breadcrumbs: Breadcrumb[] = [
+  it('should redact the access_token from the breadcrumb and request URLs', async () => {
+    const event = Event.create(
+      new Error('some error'),
+      false,
+      {
+        severity: 'error',
+        unhandled: false,
+        severityReason: {
+          type: 'blah',
+          blah: 'blah'
+        }
+      },
+      'some component',
+      3
+    );
+
+    event.breadcrumbs = [
       {
         type: 'navigation',
-        metaData: {
-          from: '/somewhere&access_token=thing',
-          to: '/projects'
-        }
+        metadata: {
+          from: 'http://localhost:5000/somewhere&access_token=thing',
+          to: 'http://localhost:5000/somewhere'
+        },
+        message: '',
+        timestamp: new Date()
       },
       {
         type: 'navigation',
-        metaData: {
-          from: '/projects#access_token=secret',
-          to: '/'
-        }
+        metadata: {
+          from: 'http://localhost:5000/projects#access_token=secret',
+          to: 'http://localhost:5000/projects'
+        },
+        message: '',
+        timestamp: new Date()
       }
     ];
+    event.request = { url: 'http://localhost:5000/projects#access_token=12345' };
 
-    const report = { breadcrumbs };
-
-    ErrorReportingService.beforeSend(report);
-    expect(report.breadcrumbs[0].metaData.from).toEqual('/somewhere&access_token=thing');
-    expect(report.breadcrumbs[0].metaData.to).toEqual('/projects');
-    expect(report.breadcrumbs[1].metaData.from).toEqual('/projects#access_token=redacted_for_error_report');
-    expect(report.breadcrumbs[1].metaData.to).toEqual('/');
+    ErrorReportingService.beforeSend({ user: null, eventId: 'eventId', locale: 'en' }, event);
+    expect(event.breadcrumbs[0].metadata.from).toEqual('http://localhost:5000/somewhere&access_token=thing');
+    expect(event.breadcrumbs[0].metadata.to).toEqual('http://localhost:5000/somewhere');
+    expect(event.breadcrumbs[1].metadata.from).toEqual(
+      'http://localhost:5000/projects#access_token=redacted_for_error_report'
+    );
+    expect(event.breadcrumbs[1].metadata.to).toEqual('http://localhost:5000/projects');
+    expect(event.request.url).toEqual('http://localhost:5000/projects#access_token=redacted_for_error_report');
   });
 });
