@@ -1,12 +1,13 @@
 import { MdcDialog, MdcDialogRef } from '@angular-mdc/web/dialog';
 import { fakeAsync, TestBed, tick } from '@angular/core/testing';
+import { NotifiableError } from '@bugsnag/js';
 import { CookieService } from 'ngx-cookie-service';
 import { User } from 'realtime-server/lib/common/models/user';
 import { Observable } from 'rxjs';
 import { anything, mock, verify, when } from 'ts-mockito';
 import { AuthService } from './auth.service';
 import { CONSOLE } from './browser-globals';
-import { ErrorReportingService } from './error-reporting.service';
+import { ErrorReportingService, EventOptions } from './error-reporting.service';
 import { ErrorComponent } from './error/error.component';
 import { ExceptionHandlingService } from './exception-handling-service';
 import { UserDoc } from './models/user-doc';
@@ -89,8 +90,8 @@ describe('ExceptionHandlingService', () => {
 
     expect(env.oneAndOnlyReport.error.message).toBe('Original error');
     expect(env.oneAndOnlyReport.error.name).toBe('Original error name');
-    expect(env.oneAndOnlyReport.opts.user.id).toBe('some id');
-    expect(env.oneAndOnlyReport.opts.metaData.eventId).toMatch(/[\da-f]{24}/);
+    expect(env.oneAndOnlyReport.options.user.id).toBe('some id');
+    expect(env.oneAndOnlyReport.options.eventId).toMatch(/[\da-f]{24}/);
   });
 
   it('should handle arbitrary objects', async () => {
@@ -119,8 +120,7 @@ describe('ExceptionHandlingService', () => {
     });
 
     expect(env.oneAndOnlyReport.error).toBeDefined();
-    expect(env.oneAndOnlyReport.opts.metaData).toBeDefined();
-    expect(env.oneAndOnlyReport.opts.user).toBeUndefined();
+    expect(env.oneAndOnlyReport.options.user).toBeUndefined();
   });
 
   it('should handle user object being unavailable', fakeAsync(() => {
@@ -129,8 +129,7 @@ describe('ExceptionHandlingService', () => {
     env.service.handleError(new Error('Test error'));
     tick(3000); // 3000ms is the time the exception handler waits for the user doc before timing out
     expect(env.oneAndOnlyReport.error).toBeDefined();
-    expect(env.oneAndOnlyReport.opts).toBeDefined();
-    expect(env.oneAndOnlyReport.opts.user).toBeUndefined();
+    expect(env.oneAndOnlyReport.options.user).toBeUndefined();
   }));
 
   it('should handle promise for user being rejected', async () => {
@@ -139,19 +138,19 @@ describe('ExceptionHandlingService', () => {
     await env.service.handleError(new Error('Test error'));
 
     expect(env.oneAndOnlyReport.error).toBeDefined();
-    expect(env.oneAndOnlyReport.opts).toBeDefined();
-    expect(env.oneAndOnlyReport.opts.user).toBeUndefined();
+    expect(env.oneAndOnlyReport.options.user).toBeUndefined();
   });
 
   it('should handle storage quota exceeded errors', async () => {
     const env = new TestEnvironment();
     await env.service.handleError(new DOMException('error', 'QuotaExceededError'));
     verify(mockedNoticeService.showError(anything())).once();
+    expect().nothing();
   });
 });
 
 class TestEnvironment {
-  readonly errorReports: { error: any; opts: any; callback: any }[] = [];
+  readonly errorReports: { error: any; options: EventOptions }[] = [];
   readonly service: ExceptionHandlingService;
   rejectUser = false;
   timeoutUser = false;
@@ -187,15 +186,9 @@ class TestEnvironment {
       });
     });
 
-    when(mockedErrorReportingService.notify(anything(), anything(), anything())).thenCall(
-      (error: any, opts: any, callback: any) => {
-        this.errorReports.push({
-          error,
-          opts,
-          callback
-        });
-      }
-    );
+    when(
+      mockedErrorReportingService.notify(anything(), anything(), anything())
+    ).thenCall((error: NotifiableError, options: EventOptions) => this.errorReports.push({ error, options }));
   }
 
   get oneAndOnlyReport() {
