@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using SIL.XForge.Models;
@@ -8,15 +9,19 @@ namespace SIL.XForge.Scripture.Services
 {
     public interface IParatextService
     {
+        void Init();
         Task<IReadOnlyList<ParatextProject>> GetProjectsAsync(UserSecret userSecret);
         string GetParatextUsername(UserSecret userSecret);
         Task<Attempt<string>> TryGetProjectRoleAsync(UserSecret userSecret, string paratextId);
-        Task<IReadOnlyList<string>> GetBooksAsync(UserSecret userSecret, string projectId);
-        Task<string> GetBookTextAsync(UserSecret userSecret, string projectId, string bookId);
-        Task<string> UpdateBookTextAsync(UserSecret userSecret, string projectId, string bookId,
-            string revision, string usxText);
-        Task<string> GetNotesAsync(UserSecret userSecret, string projectId, string bookId);
-        Task<string> UpdateNotesAsync(UserSecret userSecret, string projectId, string notesText);
         Task<IReadOnlyDictionary<string, string>> GetProjectRolesAsync(UserSecret userSecret, string projectId);
+
+        IReadOnlyList<int> GetBookList(UserSecret userSecret, string ptProjectId, TextType textType);
+        string GetBookText(UserSecret userSecret, string ptProjectId, int bookNum, TextType textType);
+        void PutBookText(UserSecret userSecret, string ptProjectId, int bookNum, string usx);
+        string GetNotes(UserSecret userSecret, string ptProjectId, int bookNum);
+        void PutNotes(UserSecret userSecret, string ptProjectId, string notesText);
+
+        Task SendReceiveAsync(UserSecret userSecret, string ptTargetId, string ptSourceId,
+            IProgress<ProgressState> progress = null);
     }
 }
