@@ -1,135 +1,224 @@
 using System;
 using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
+using System.Reflection;
+using System.Runtime.InteropServices;
 using System.Security;
 using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
-using System.Xml.Linq;
+using System.Xml;
+using System.Xml.XPath;
 using IdentityModel;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json.Linq;
+using Paratext.Data;
+using Paratext.Data.Languages;
+using Paratext.Data.RegistryServerAccess;
+using Paratext.Data.Repository;
+using Paratext.Data.Users;
+using Paratext.Data.ProjectComments;
+using PtxUtils;
 using SIL.ObjectModel;
 using SIL.XForge.Configuration;
 using SIL.XForge.DataAccess;
 using SIL.XForge.Models;
 using SIL.XForge.Realtime;
 using SIL.XForge.Scripture.Models;
+using SIL.XForge.Services;
 using SIL.XForge.Utils;
+using SIL.Scripture;
+using System.Diagnostics;
 
 namespace SIL.XForge.Scripture.Services
 {
     /// <summary>
-    /// This class contains methods for interacting with the Paratext web service APIs.
+    /// Provides interaction with Paratext libraries for data processing and exchanging data with Paratext servers.
+    /// Also contains methods for interacting with the Paratext Registry web service API.
     /// </summary>
     public class ParatextService : DisposableBase, IParatextService
     {
-        private readonly IOptions<ParatextOptions> _options;
-        private readonly IRepository<UserSecret> _userSecret;
+        private readonly IWebHostEnvironment _webHostEnvironment;
+        private readonly IOptions<ParatextOptions> _paratextOptions;
+        private readonly IRepository<UserSecret> _userSecretRepository;
         private readonly IRealtimeService _realtimeService;
+        private readonly IOptions<SiteOptions> _siteOptions;
+        private readonly IFileSystemService _fileSystemService;
         private readonly HttpClientHandler _httpClientHandler;
-        private readonly HttpClient _dataAccessClient;
         private readonly HttpClient _registryClient;
         private readonly IExceptionHandler _exceptionHandler;
+        private readonly ILogger _logger;
+        private readonly IJwtTokenHelper _jwtTokenHelper;
+        private string _applicationProductVersion = "SF";
+        private string _registryServerUri = "https://registry.paratext.org";
+        private string _sendReceiveServerUri = InternetAccess.uriProduction;
+        private readonly IInternetSharedRepositorySourceProvider _internetSharedRepositorySourceProvider;
 
-        public ParatextService(IWebHostEnvironment env, IOptions<ParatextOptions> options,
-            IRepository<UserSecret> userSecret, IRealtimeService realtimeService, IExceptionHandler exceptionHandler)
+
+        public ParatextService(IWebHostEnvironment env, IOptions<ParatextOptions> paratextOptions,
+            IRepository<UserSecret> userSecretRepository, IRealtimeService realtimeService,
+            IExceptionHandler exceptionHandler, IOptions<SiteOptions> siteOptions, IFileSystemService fileSystemService,
+            ILogger<ParatextService> logger, IJwtTokenHelper jwtTokenHelper,
+            IInternetSharedRepositorySourceProvider internetSharedRepositorySourceProvider)
         {
-            _options = options;
-            _userSecret = userSecret;
+            _webHostEnvironment = env;
+            _paratextOptions = paratextOptions;
+            _userSecretRepository = userSecretRepository;
             _realtimeService = realtimeService;
             _exceptionHandler = exceptionHandler;
+            _siteOptions = siteOptions;
+            _fileSystemService = fileSystemService;
+            _logger = logger;
+            _jwtTokenHelper = jwtTokenHelper;
+            _internetSharedRepositorySourceProvider = internetSharedRepositorySourceProvider;
 
             _httpClientHandler = new HttpClientHandler();
-            _dataAccessClient = new HttpClient(_httpClientHandler);
             _registryClient = new HttpClient(_httpClientHandler);
             if (env.IsDevelopment() || env.IsEnvironment("Testing"))
             {
                 _httpClientHandler.ServerCertificateCustomValidationCallback
                     = HttpClientHandler.DangerousAcceptAnyServerCertificateValidator;
-                _dataAccessClient.BaseAddress = new Uri("https://data-access-dev.paratext.org");
-                _registryClient.BaseAddress = new Uri("https://registry-dev.paratext.org");
+                _registryServerUri = "https://registry-dev.paratext.org";
+                _registryClient.BaseAddress = new Uri(_registryServerUri);
+                _sendReceiveServerUri = InternetAccess.uriDevelopment;
             }
             else
             {
-                _dataAccessClient.BaseAddress = new Uri("https://data-access.paratext.org");
-                _registryClient.BaseAddress = new Uri("https://registry.paratext.org");
+                _registryClient.BaseAddress = new Uri(_registryServerUri);
             }
             _registryClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+            ScrTextCollection = new LazyScrTextCollection();
+            HgWrapper = new HgWrapper();
+
+            SharingLogicWrapper = new SharingLogicWrapper();
+            Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
         }
 
+        public static string AssemblyDirectory
+        {
+            get
+            {
+                string codeBase = Assembly.GetExecutingAssembly().CodeBase;
+                UriBuilder uri = new UriBuilder(codeBase);
+                string path = Uri.UnescapeDataString(uri.Path);
+                return Path.GetDirectoryName(path);
+            }
+        }
+
+        ///< summary> Path to cloned PT project Mercurial repos. </summary>
+        public string SyncDir { get; set; }
+
+        internal IScrTextCollection ScrTextCollection { get; set; }
+        internal ISharingLogicWrapper SharingLogicWrapper { get; set; }
+        internal IHgWrapper HgWrapper { get; set; }
+
+        /// <summary> Prepare access to Paratext.Data library, authenticate, and prepare Mercurial. </summary>
+        public void Init()
+        {
+            if (_webHostEnvironment.IsDevelopment() || _webHostEnvironment.IsEnvironment("Testing"))
+            {
+                // On dev machines, output more info from ParatextData.dll for investigating.
+                Trace.Listeners.Add(new TextWriterTraceListener(Console.Out));
+                Trace.AutoFlush = true;
+            }
+
+            SyncDir = Path.Combine(_siteOptions.Value.SiteDir, "sync");
+            if (!_fileSystemService.DirectoryExists(SyncDir))
+                _fileSystemService.CreateDirectory(SyncDir);
+            // Disable caching VersionedText instances since multiple repos may exist on SF server with the same GUID
+            Environment.SetEnvironmentVariable("PTD_CACHE_VERSIONED_TEXT", "DISABLED");
+            RegistryU.Implementation = new DotNetCoreRegistry();
+            Alert.Implementation = new DotNetCoreAlert(_logger);
+            ParatextDataSettings.Initialize(new PersistedParatextDataSettings());
+            PtxUtilsDataSettings.Initialize(new PersistedPtxUtilsSettings());
+            SetupMercurial();
+            WritingSystemRepository.Initialize();
+            ScrTextCollection.Initialize(SyncDir);
+            RegistryServer.Initialize(_applicationProductVersion);
+            InstallStyles();
+            // Allow use of custom versification systems
+            Versification.Table.Implementation = new ParatextVersificationTable();
+        }
+
+        /// <summary>
+        /// Synchronizes the text and notes data on the SF server with the data on the Paratext server.
+        /// </summary>
+        public async Task SendReceiveAsync(UserSecret userSecret, string ptTargetId, string ptSourceId,
+            IProgress<ProgressState> progress = null)
+        {
+            if (userSecret == null || ptTargetId == null) { throw new ArgumentNullException(); }
+
+            IInternetSharedRepositorySource source = await GetInternetSharedRepositorySource(userSecret);
+            IEnumerable<SharedRepository> repositories = source.GetRepositories();
+            IEnumerable<ProjectMetadata> projectsMetadata = source.GetProjectsMetaData();
+            var projectGuids = projectsMetadata.Select(pmd => pmd.ProjectGuid);
+            Dictionary<string, ParatextProject> ptProjectsAvailable =
+                GetProjects(userSecret, repositories, projectsMetadata).ToDictionary(ptProject => ptProject.ParatextId);
+            if (!projectGuids.Contains(ptTargetId))
+            {
+                _logger.LogWarning($"The target project did not have a full name available {ptTargetId}");
+            }
+            if (!projectGuids.Contains(ptSourceId) && ptSourceId != null)
+            {
+                _logger.LogWarning($"The source project did not have a full name available {ptSourceId}");
+            }
+            List<string> problemProjectIds = new List<string>();
+            if (!ptProjectsAvailable.TryGetValue(ptTargetId, out ParatextProject targetPtProject))
+                problemProjectIds.Add(ptTargetId);
+            ParatextProject sourcePtProject = null;
+            if (ptSourceId != null && !ptProjectsAvailable.TryGetValue(ptSourceId, out sourcePtProject))
+                problemProjectIds.Add(ptSourceId);
+            if (problemProjectIds.Count > 0)
+            {
+                throw new ArgumentException(
+                    "PT projects with the following PT ids were requested but without access or they don't exist: " +
+                        string.Join(", ", problemProjectIds));
+            }
+
+            EnsureProjectReposExists(userSecret, targetPtProject, sourcePtProject, source);
+            StartProgressReporting(progress);
+            SharedProject targetSharedProj = SharingLogicWrapper.CreateSharedProject(ptTargetId,
+                targetPtProject.ShortName, source.AsInternetSharedRepositorySource(), repositories);
+            string username = GetParatextUsername(userSecret);
+            // Specifically set the ScrText property of the SharedProject to indicate the project is available locally
+            targetSharedProj.ScrText = ScrTextCollection.FindById(username, ptTargetId, Models.TextType.Target);
+            targetSharedProj.Permissions = targetSharedProj.ScrText.Permissions;
+            List<SharedProject> sharedPtProjectsToSr = new List<SharedProject> { targetSharedProj };
+            if (sourcePtProject != null)
+            {
+                SharedProject sourceSharedProj = SharingLogicWrapper.CreateSharedProject(ptSourceId,
+                    sourcePtProject.ShortName, source.AsInternetSharedRepositorySource(), repositories);
+                sourceSharedProj.ScrText = ScrTextCollection.FindById(username, ptTargetId, Models.TextType.Source);
+                sourceSharedProj.Permissions = sourceSharedProj.ScrText.Permissions;
+                sharedPtProjectsToSr.Add(sourceSharedProj);
+            }
+
+            // TODO report results
+            List<SendReceiveResult> results = Enumerable.Empty<SendReceiveResult>().ToList();
+            bool success = false;
+            bool noErrors = SharingLogicWrapper.HandleErrors(() => success = SharingLogicWrapper
+                .ShareChanges(sharedPtProjectsToSr, source.AsInternetSharedRepositorySource(),
+                out results, sharedPtProjectsToSr));
+            if (!noErrors || !success)
+                throw new InvalidOperationException(
+                    "Failed: Errors occurred while performing the sync with the Paratext Server.");
+        }
+
+        /// <summary> Get Paratext projects that a user has access to. </summary>
         public async Task<IReadOnlyList<ParatextProject>> GetProjectsAsync(UserSecret userSecret)
         {
-            var accessToken = new JwtSecurityToken(userSecret.ParatextTokens.AccessToken);
-            Claim usernameClaim = accessToken.Claims.FirstOrDefault(c => c.Type == "username");
-            string username = usernameClaim?.Value;
-            string response = await CallApiAsync(_dataAccessClient, userSecret, HttpMethod.Get, "projects");
-            var reposElem = XElement.Parse(response);
-            var repos = new Dictionary<string, string>();
-            foreach (XElement repoElem in reposElem.Elements("repo"))
-            {
-                var projId = (string)repoElem.Element("projid");
-                XElement userElem = repoElem.Element("users")?.Elements("user")
-                    ?.FirstOrDefault(ue => (string)ue.Element("name") == username);
-                repos[projId] = (string)userElem?.Element("role");
-            }
-            Dictionary<string, SFProject> existingProjects = (await _realtimeService.QuerySnapshots<SFProject>()
-                .Where(p => repos.Keys.Contains(p.ParatextId))
-                .ToListAsync()).ToDictionary(p => p.ParatextId);
-            response = await CallApiAsync(_registryClient, userSecret, HttpMethod.Get, "projects");
-            var projectArray = JArray.Parse(response);
-            var projects = new List<ParatextProject>();
-            foreach (JToken projectObj in projectArray)
-            {
-                JToken identificationObj = projectObj["identification_systemId"]
-                    .FirstOrDefault(id => (string)id["type"] == "paratext");
-                if (identificationObj == null)
-                    continue;
-                string paratextId = (string)identificationObj["text"];
-                if (!repos.TryGetValue(paratextId, out string role))
-                    continue;
-
-                // determine if the project is connectable, i.e. either the project exists and the user hasn't been
-                // added to the project, or the project doesn't exist and the user is the administrator
-                bool isConnectable;
-                bool isConnected = false;
-                string projectId = null;
-                if (existingProjects.TryGetValue(paratextId, out SFProject project))
-                {
-                    projectId = project.Id;
-                    isConnected = true;
-                    isConnectable = !project.UserRoles.ContainsKey(userSecret.Id);
-                }
-                else if (role == SFProjectRole.Administrator)
-                {
-                    isConnectable = true;
-                }
-                else
-                {
-                    isConnectable = false;
-                }
-
-                projects.Add(new ParatextProject
-                {
-                    ParatextId = paratextId,
-                    Name = (string)projectObj["identification_name"],
-                    ShortName = (string)projectObj["identification_shortName"],
-                    LanguageTag = (string)projectObj["language_ldml"],
-                    ProjectId = projectId,
-                    IsConnectable = isConnectable,
-                    IsConnected = isConnected
-                });
-            }
-            return projects.OrderBy(p => p.Name, StringComparer.InvariantCulture).ToArray();
+            IInternetSharedRepositorySource ptRepoSource = await GetInternetSharedRepositorySource(userSecret);
+            IEnumerable<SharedRepository> remotePtProjects = ptRepoSource.GetRepositories();
+            return GetProjects(userSecret, remotePtProjects, ptRepoSource.GetProjectsMetaData());
         }
-
-
 
         public async Task<Attempt<string>> TryGetProjectRoleAsync(UserSecret userSecret, string paratextId)
         {
@@ -139,6 +228,8 @@ namespace SIL.XForge.Scripture.Services
             {
                 var accessToken = new JwtSecurityToken(userSecret.ParatextTokens.AccessToken);
                 Claim subClaim = accessToken.Claims.FirstOrDefault(c => c.Type == JwtClaimTypes.Subject);
+                // Paratext RegistryServer has methods to do this, but it is unreliable to use it in a multi-user
+                // environment so instead we call the registry API.
                 string response = await CallApiAsync(_registryClient, userSecret, HttpMethod.Get,
                     $"projects/{paratextId}/members/{subClaim.Value}");
                 var memberObj = JObject.Parse(response);
@@ -150,49 +241,17 @@ namespace SIL.XForge.Scripture.Services
             }
         }
 
+        /// <summary> Get the Paratext username from the UserSecret. </summary>
         public string GetParatextUsername(UserSecret userSecret)
         {
-            if (userSecret.ParatextTokens == null)
-                return null;
-            var accessToken = new JwtSecurityToken(userSecret.ParatextTokens.AccessToken);
-            Claim usernameClaim = accessToken.Claims.FirstOrDefault(c => c.Type == "username");
-            return usernameClaim?.Value;
-        }
-
-        public async Task<IReadOnlyList<string>> GetBooksAsync(UserSecret userSecret, string projectId)
-        {
-            string response = await CallApiAsync(_dataAccessClient, userSecret, HttpMethod.Get, $"books/{projectId}");
-            var books = XElement.Parse(response);
-            string[] bookIds = books.Elements("Book").Select(b => (string)b.Attribute("id")).ToArray();
-            return bookIds;
-        }
-
-        public Task<string> GetBookTextAsync(UserSecret userSecret, string projectId, string bookId)
-        {
-            return CallApiAsync(_dataAccessClient, userSecret, HttpMethod.Get, $"text/{projectId}/{bookId}");
-        }
-
-        /// <summary>Update cloud with new edits in usxText and return the combined result.</summary>
-        public Task<string> UpdateBookTextAsync(UserSecret userSecret, string projectId, string bookId,
-            string revision, string usxText)
-        {
-            return CallApiAsync(_dataAccessClient, userSecret, HttpMethod.Post,
-                $"text/{projectId}/{revision}/{bookId}", usxText);
-        }
-
-        public Task<string> GetNotesAsync(UserSecret userSecret, string projectId, string bookId)
-        {
-            return CallApiAsync(_dataAccessClient, userSecret, HttpMethod.Get, $"notes/{projectId}/{bookId}");
-        }
-
-        public Task<string> UpdateNotesAsync(UserSecret userSecret, string projectId, string notesText)
-        {
-            return CallApiAsync(_dataAccessClient, userSecret, HttpMethod.Post, $"notes/{projectId}", notesText);
+            return _jwtTokenHelper.GetParatextUsername(userSecret);
         }
 
         public async Task<IReadOnlyDictionary<string, string>> GetProjectRolesAsync(UserSecret userSecret,
             string projectId)
         {
+            // Paratext RegistryServer has methods to do this, but it is unreliable to use it in a multi-user
+            // environment so instead we call the registry API.
             string response = await CallApiAsync(_registryClient, userSecret, HttpMethod.Get,
                 $"projects/{projectId}/members");
             var members = JArray.Parse(response);
@@ -201,28 +260,254 @@ namespace SIL.XForge.Scripture.Services
                 .ToDictionary(m => (string)m["userId"], m => (string)m["role"]);
         }
 
+        /// <summary> Get list of book numbers in PT project. </summary>
+        public IReadOnlyList<int> GetBookList(UserSecret userSecret, string ptProjectId, Models.TextType textType)
+        {
+            ScrText scrText = ScrTextCollection.FindById(GetParatextUsername(userSecret), ptProjectId, textType);
+            if (scrText == null)
+                return Array.Empty<int>();
+            return scrText.Settings.BooksPresentSet.SelectedBookNumbers.ToArray();
+        }
+
+        /// <summary> Get PT book text in USX, or throw if can't. </summary>
+        public string GetBookText(UserSecret userSecret, string ptProjectId, int bookNum, Models.TextType textType)
+        {
+            ScrText scrText = ScrTextCollection.FindById(GetParatextUsername(userSecret), ptProjectId, textType);
+            if (scrText == null)
+                throw new DataNotFoundException("Can't get access to cloned project.");
+            string usfm = scrText.GetText(bookNum);
+            return UsfmToUsx.ConvertToXmlString(scrText, bookNum, usfm, false);
+        }
+
+        /// <summary> Write up-to-date book text from mongo database to Paratext project folder. </summary>
+        public void PutBookText(UserSecret userSecret, string projectId, int bookNum, string usx)
+        {
+            ScrText scrText = ScrTextCollection.FindById(GetParatextUsername(userSecret), projectId,
+                Models.TextType.Target);
+            var doc = new XmlDocument
+            {
+                PreserveWhitespace = true
+            };
+            doc.LoadXml(usx);
+            UsxFragmenter.FindFragments(scrText.ScrStylesheet(bookNum), doc.CreateNavigator(),
+                XPathExpression.Compile("*[false()]"), out string usfm);
+            usfm = UsfmToken.NormalizeUsfm(scrText.ScrStylesheet(bookNum), usfm, false, scrText.RightToLeft);
+            scrText.PutText(bookNum, 0, false, usfm, null);
+            _logger.LogInformation("{0} updated {1} in {2}.", GetParatextUsername(userSecret),
+                Canon.BookNumberToEnglishName(bookNum), scrText.Name);
+        }
+
+        /// <summary> Get notes from the Paratext project folder. </summary>
+        public string GetNotes(UserSecret userSecret, string projectId, int bookNum)
+        {
+            // TODO: should return some data structure instead of XML
+            ScrText scrText = ScrTextCollection.FindById(GetParatextUsername(userSecret), projectId,
+                Models.TextType.Target);
+            if (scrText == null)
+                return null;
+
+            CommentManager manager = CommentManager.Get(scrText);
+            var threads = manager.FindThreads((commentThread) => { return commentThread.VerseRef.BookNum == bookNum; },
+                true);
+            return NotesFormatter.FormatNotes(threads);
+        }
+
+        /// <summary> Write up-to-date notes from the mongo database to the Paratext project folder </summary>
+        public void PutNotes(UserSecret userSecret, string projectId, string notesText)
+        {
+            // TODO: should accept some data structure instead of XML
+            string username = GetParatextUsername(userSecret);
+            List<string> users = new List<string>();
+            int nbrAddedComments = 0, nbrDeletedComments = 0, nbrUpdatedComments = 0;
+            ScrText scrText = ScrTextCollection.FindById(username, projectId, Models.TextType.Target);
+            if (scrText == null)
+                throw new DataNotFoundException("Can't get access to cloned project.");
+            CommentManager manager = CommentManager.Get(scrText);
+            var notes = NotesFormatter.ParseNotes(notesText);
+
+            // Algorithm sourced from Paratext DataAccessServer
+            foreach (var thread in notes)
+            {
+                CommentThread existingThread = manager.FindThread(thread[0].Thread);
+                foreach (var comment in thread)
+                {
+                    var existingComment = existingThread?.Comments.FirstOrDefault(c => c.Id == comment.Id);
+                    if (existingComment == null)
+                    {
+                        manager.AddComment(comment);
+                        nbrAddedComments++;
+                    }
+                    else if (comment.Deleted)
+                    {
+                        existingComment.Deleted = true;
+                        nbrDeletedComments++;
+                    }
+                    else
+                    {
+                        existingComment.ExternalUser = comment.ExternalUser;
+                        existingComment.Contents = comment.Contents;
+                        existingComment.VersionNumber += 1;
+                        nbrUpdatedComments++;
+                    }
+
+                    if (!users.Contains(comment.User))
+                        users.Add(comment.User);
+                }
+            }
+
+            try
+            {
+                foreach (string user in users)
+                    manager.SaveUser(user, false);
+                VersionedText vText = VersioningManager.Get(scrText);
+                vText.Commit($"{nbrAddedComments} notes added and {nbrDeletedComments + nbrUpdatedComments} notes updated or deleted in synchronize",
+                     null, false, username);
+                _logger.LogInformation("{0} added {1} notes, updated {2} notes and deleted {3} notes",
+                    username, nbrAddedComments, nbrUpdatedComments, nbrDeletedComments);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e, "Exception while updating notes: {0}", e.Message);
+            }
+        }
+
+        protected override void DisposeManagedResources()
+        {
+            _registryClient.Dispose();
+            _httpClientHandler.Dispose();
+        }
+
+        private IReadOnlyList<ParatextProject> GetProjects(UserSecret userSecret,
+            IEnumerable<SharedRepository> remotePtProjects, IEnumerable<ProjectMetadata> projectsMetadata)
+        {
+            if (userSecret == null) throw new ArgumentNullException();
+
+            List<ParatextProject> paratextProjects = new List<ParatextProject>();
+            IQueryable<SFProject> existingSfProjects = _realtimeService.QuerySnapshots<SFProject>();
+
+            foreach (SharedRepository remotePtProject in remotePtProjects)
+            {
+                SFProject correspondingSfProject =
+                    existingSfProjects.FirstOrDefault(sfProj => sfProj.ParatextId == remotePtProject.SendReceiveId);
+
+                bool sfProjectExists = correspondingSfProject != null;
+                bool sfUserIsOnSfProject = correspondingSfProject?.UserRoles.ContainsKey(userSecret.Id) ?? false;
+                bool adminOnPtProject = remotePtProject.SourceUsers.GetRole(
+                    GetParatextUsername(userSecret)) == UserRoles.Administrator;
+                bool ptProjectIsConnectable =
+                    (sfProjectExists && !sfUserIsOnSfProject) || (!sfProjectExists && adminOnPtProject);
+
+                // On SF Live server, many users have projects without corresponding project metadata.
+                // If this happens, default to using the project's short name
+                var projectMD = projectsMetadata
+                    .SingleOrDefault(pmd => pmd.ProjectGuid == remotePtProject.SendReceiveId);
+                string fullOrShortName = projectMD == null ? remotePtProject.ScrTextName : projectMD.FullName;
+
+                paratextProjects.Add(new ParatextProject
+                {
+                    ParatextId = remotePtProject.SendReceiveId,
+                    Name = fullOrShortName,
+                    ShortName = remotePtProject.ScrTextName,
+                    LanguageTag = correspondingSfProject?.WritingSystem.Tag,
+                    ProjectId = correspondingSfProject?.Id,
+                    IsConnectable = ptProjectIsConnectable,
+                    IsConnected = sfProjectExists && sfUserIsOnSfProject
+                });
+            }
+            return paratextProjects.OrderBy(project => project.Name, StringComparer.InvariantCulture).ToArray();
+        }
+
+        private void SetupMercurial()
+        {
+            // We do not yet know where hg will be installed on the server, so allow defining it in an env variable
+            string customHgPath = Environment.GetEnvironmentVariable("HG_PATH") ?? _paratextOptions.Value.HgExe;
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                customHgPath = Path.GetExtension(customHgPath) != ".exe" ? customHgPath + ".exe" : customHgPath;
+            if (!File.Exists(customHgPath))
+            {
+                string msg = string.Format(
+                    "Error: Could not find hg executable at {0}. Please install hg 4.7 or greater.", customHgPath);
+                _logger.LogError(msg);
+                throw new InvalidOperationException(msg);
+            }
+            var hgMerge = Path.Combine(AssemblyDirectory, "ParatextMerge.py");
+            HgWrapper.SetDefault(new Hg(customHgPath, hgMerge, AssemblyDirectory));
+        }
+
+        /// <summary> Copy resource files from the Assembly Directory into the sync directory. </summary>
+        private void InstallStyles()
+        {
+            string[] resources = new[] { "usfm.sty", "revisionStyle.sty", "revisionTemplate.tem", "usfm_mod.sty" };
+            foreach (string resource in resources)
+            {
+                string target = Path.Combine(SyncDir, resource);
+                string source = Path.Combine(AssemblyDirectory, resource);
+                if (!File.Exists(target))
+                {
+                    _logger.LogInformation($"Installing missing {target}");
+                    File.Copy(source, target, true);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Ensure the target and source project repositories exist on the local SF server, cloning them if necessary.
+        /// </summary>
+        private void EnsureProjectReposExists(UserSecret userSecret, ParatextProject target, ParatextProject source,
+            IInternetSharedRepositorySource repositorySource)
+        {
+            string username = GetParatextUsername(userSecret);
+            bool targetNeedsCloned =
+                ScrTextCollection.FindById(username, target.ParatextId, Models.TextType.Target) == null;
+            if (targetNeedsCloned)
+            {
+                SharedRepository targetRepo = new SharedRepository(target.ShortName, target.ParatextId,
+                    RepositoryType.Shared);
+                CloneProjectRepo(repositorySource, target.ParatextId, targetRepo, Models.TextType.Target);
+            }
+            if (source != null)
+            {
+                bool sourceNeedsCloned = false;
+                ScrText sourceScrText = ScrTextCollection.FindById(username, target.ParatextId, Models.TextType.Source);
+                if (sourceScrText == null)
+                    sourceNeedsCloned = true;
+                else if (sourceScrText.Guid != source.ParatextId)
+                {
+                    // The source project has changed. So delete the obsolete source project
+                    _fileSystemService.DeleteDirectory(sourceScrText.Directory);
+                    sourceNeedsCloned = true;
+                }
+                if (sourceNeedsCloned)
+                {
+                    SharedRepository sourceRepo = new SharedRepository(source.ShortName, source.ParatextId,
+                        RepositoryType.Shared);
+                    CloneProjectRepo(repositorySource, target.ParatextId, sourceRepo, Models.TextType.Source);
+                }
+            }
+        }
+
+        private void CloneProjectRepo(IInternetSharedRepositorySource source, string projectId, SharedRepository repo,
+            Models.TextType textType)
+        {
+            string clonePath = Path.Combine(SyncDir, projectId, TextTypeUtils.DirectoryName(textType));
+            if (!_fileSystemService.DirectoryExists(clonePath))
+            {
+                _fileSystemService.CreateDirectory(clonePath);
+                HgWrapper.Init(clonePath);
+            }
+            source.Pull(clonePath, repo);
+            HgWrapper.Update(clonePath);
+        }
+
         private async Task RefreshAccessTokenAsync(UserSecret userSecret)
         {
-            var request = new HttpRequestMessage(HttpMethod.Post, "api8/token");
+            ParatextOptions options = _paratextOptions.Value;
 
-            ParatextOptions options = _options.Value;
-            var requestObj = new JObject(
-                new JProperty("grant_type", "refresh_token"),
-                new JProperty("client_id", options.ClientId),
-                new JProperty("client_secret", options.ClientSecret),
-                new JProperty("refresh_token", userSecret.ParatextTokens.RefreshToken));
-            request.Content = new StringContent(requestObj.ToString(), Encoding.UTF8, "application/json");
-            HttpResponseMessage response = await _registryClient.SendAsync(request);
-            await _exceptionHandler.EnsureSuccessStatusCode(response);
+            userSecret.ParatextTokens = await _jwtTokenHelper.RefreshAccessTokenAsync(options,
+                userSecret.ParatextTokens, _registryClient);
 
-            string responseJson = await response.Content.ReadAsStringAsync();
-            var responseObj = JObject.Parse(responseJson);
-            userSecret.ParatextTokens = new Tokens
-            {
-                AccessToken = (string)responseObj["access_token"],
-                RefreshToken = (string)responseObj["refresh_token"],
-            };
-            await _userSecret.UpdateAsync(userSecret, b => b.Set(u => u.ParatextTokens, userSecret.ParatextTokens));
+            await _userSecretRepository.UpdateAsync(userSecret, b =>
+                b.Set(u => u.ParatextTokens, userSecret.ParatextTokens));
         }
 
         private async Task<string> CallApiAsync(HttpClient client, UserSecret userSecret, HttpMethod method,
@@ -264,11 +549,25 @@ namespace SIL.XForge.Scripture.Services
             throw new SecurityException("The current user's Paratext access token is invalid.");
         }
 
-        protected override void DisposeManagedResources()
+        /// <summary>
+        /// Get access to a source for PT project repositories, based on user secret.
+        ///</summary>
+        private async Task<IInternetSharedRepositorySource> GetInternetSharedRepositorySource(UserSecret userSecret)
         {
-            _dataAccessClient.Dispose();
-            _registryClient.Dispose();
-            _httpClientHandler.Dispose();
+            if (userSecret == null) throw new ArgumentNullException();
+            await RefreshAccessTokenAsync(userSecret);
+            IInternetSharedRepositorySource source = _internetSharedRepositorySourceProvider.GetSource(userSecret,
+                _sendReceiveServerUri, _registryServerUri, _applicationProductVersion);
+            return source;
+        }
+
+        // Make sure there are no asynchronous methods called after this until the progress is completed.
+        private void StartProgressReporting(IProgress<ProgressState> progress)
+        {
+            if (progress == null)
+                return;
+            var progressDisplay = new SyncProgressDisplay(progress);
+            PtxUtils.Progress.Progress.Mgr.SetDisplay(progressDisplay);
         }
     }
 }
