@@ -1,55 +1,49 @@
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using SIL.Machine.WebApi.Services;
 using SIL.ObjectModel;
-using SIL.Scripture;
-using SIL.XForge.Configuration;
 using SIL.XForge.DataAccess;
 using SIL.XForge.Models;
 using SIL.XForge.Realtime;
 using SIL.XForge.Realtime.Json0;
 using SIL.XForge.Realtime.RichText;
 using SIL.XForge.Scripture.Models;
-using SIL.XForge.Services;
 
 namespace SIL.XForge.Scripture.Services
 {
     /// <summary>
-    /// This class syncs real-time text and question docs with the Paratext data access API.
+    /// This class syncs real-time text and question docs with the Paratext S/R server. This class coordinates the data
+    /// flow between three sources:
+    /// 1. The real-time docs stored in Mongo.
+    /// 2. The local Paratext project repo.
+    /// 3. The remote Paratext project repo.
     ///
-    /// Text sync (two-way):
+    /// Algorithm:
     /// 1. The text deltas from the real-time docs are converted to USX.
-    /// 2. The USX is sent to the PT data access API if the USX has changed since the last sync (this is determined by
-    /// checking the new USX against a stored copy of the USX from the last sync).
-    /// 3. The PT data access API merges the USX with any other changes and returns the merged USX.
-    /// 4. The returned USX is converted back to text deltas and diffed against the current deltas.
-    /// 5. The diff is submitted as an operation to the real-time text docs.
+    /// 2. The local repo is updated using the converted USX.
+    /// 3. A note changelist is computed by diffing the the real-time question docs and the notes in the local repo.
+    /// 4. The local repo is updated using the note changelist.
+    /// 5. PT send/receive is performed (remote and local repos are synced).
+    /// 6. Docs associated with remotely deleted books are deleted.
+    /// 7. Updated USX is retrieved from the local repo.
+    /// 7. The returned USX is converted back to text deltas and diffed against the current deltas.
+    /// 8. The diff is submitted as an operation to the real-time text docs (chapter docs are added and deleted as
+    /// needed).
+    /// 9. Question docs associated with remotely deleted chapters are deleted.
     ///
-    /// Notes sync (one-way):
-    /// 1. The current notes are retrieved from the PT data access API.
-    /// 2. A notes changelist XML is generated from the real-time question docs.
-    /// 3. The notes changelist is sent to the PT data access API.
-    ///
-    /// Target and source refer to child and mother translation data. Not
-    /// to be confused with a target or source for where data is coming
-    /// from or going to when fetching or syncing.
+    /// Target and source refer to daughter and mother translation data. Not to be confused with a target or source for
+    /// where data is coming from or going to when fetching or syncing.
     /// </summary>
     public class ParatextSyncRunner : IParatextSyncRunner
     {
         private static readonly IEqualityComparer<List<Chapter>> ChapterListEqualityComparer =
             SequenceEqualityComparer.Create(new ChapterEqualityComparer());
 
-        private readonly IOptions<SiteOptions> _siteOptions;
         private readonly IRepository<UserSecret> _userSecrets;
         private readonly IRepository<SFProjectSecret> _projectSecrets;
         private readonly ISFProjectService _projectService;
@@ -58,35 +52,28 @@ namespace SIL.XForge.Scripture.Services
         private readonly IRealtimeService _realtimeService;
         private readonly IDeltaUsxMapper _deltaUsxMapper;
         private readonly IParatextNotesMapper _notesMapper;
-        private readonly IFileSystemService _fileSystemService;
         private readonly ILogger<ParatextSyncRunner> _logger;
 
         private IConnection _conn;
         private UserSecret _userSecret;
         private IDocument<SFProject> _projectDoc;
         private SFProjectSecret _projectSecret;
-        private int _stepCount;
-        private int _step;
 
-        public ParatextSyncRunner(IOptions<SiteOptions> siteOptions, IRepository<UserSecret> userSecrets,
-            IRepository<SFProjectSecret> projectSecrets, ISFProjectService projectService, IEngineService engineService,
-            IParatextService paratextService, IRealtimeService realtimeService, IFileSystemService fileSystemService,
-            IDeltaUsxMapper deltaUsxMapper, IParatextNotesMapper notesMapper, ILogger<ParatextSyncRunner> logger)
+        public ParatextSyncRunner(IRepository<UserSecret> userSecrets, IRepository<SFProjectSecret> projectSecrets,
+            ISFProjectService projectService, IEngineService engineService, IParatextService paratextService,
+            IRealtimeService realtimeService, IDeltaUsxMapper deltaUsxMapper, IParatextNotesMapper notesMapper,
+            ILogger<ParatextSyncRunner> logger)
         {
-            _siteOptions = siteOptions;
             _userSecrets = userSecrets;
             _projectSecrets = projectSecrets;
             _projectService = projectService;
             _engineService = engineService;
             _paratextService = paratextService;
             _realtimeService = realtimeService;
-            _fileSystemService = fileSystemService;
             _logger = logger;
             _deltaUsxMapper = deltaUsxMapper;
             _notesMapper = notesMapper;
         }
-
-        private string WorkingDir => Path.Combine(_siteOptions.Value.SiteDir, "sync");
 
         private bool TranslationSuggestionsEnabled => _projectDoc.Data.TranslateConfig.TranslationSuggestionsEnabled;
         private bool CheckingEnabled => _projectDoc.Data.CheckingConfig.CheckingEnabled;
@@ -107,31 +94,50 @@ namespace SIL.XForge.Scripture.Services
                 string targetParatextId = _projectDoc.Data.ParatextId;
                 string sourceParatextId = _projectDoc.Data.TranslateConfig.Source?.ParatextId;
 
-                var targetBooks = new HashSet<int>((await _paratextService.GetBooksAsync(_userSecret,
-                    targetParatextId)).Select(bookId => Canon.BookIdToNumber(bookId)));
+                var targetTextDocsByBook = new Dictionary<int, SortedList<int, IDocument<TextData>>>();
+                var sourceTextDocsByBook = new Dictionary<int, SortedList<int, IDocument<TextData>>>();
+                var questionDocsByBook = new Dictionary<int, IReadOnlyList<IDocument<Question>>>();
 
+                // update target Paratext books and notes
+                foreach (TextInfo text in _projectDoc.Data.Texts)
+                {
+                    SortedList<int, IDocument<TextData>> targetTextDocs = await FetchTextDocsAsync(text,
+                        TextType.Target);
+                    targetTextDocsByBook[text.BookNum] = targetTextDocs;
+                    UpdateParatextBook(text, targetParatextId, targetTextDocs);
+                    if (text.HasSource)
+                        sourceTextDocsByBook[text.BookNum] = await FetchTextDocsAsync(text, TextType.Source);
+
+                    IReadOnlyList<IDocument<Question>> questionDocs = await FetchQuestionDocsAsync(text);
+                    questionDocsByBook[text.BookNum] = questionDocs;
+                    await UpdateParatextNotesAsync(text, questionDocs);
+                }
+
+                // perform Paratext send/receive
+                await _paratextService.SendReceiveAsync(_userSecret, targetParatextId, sourceParatextId,
+                    UseNewProgress());
+
+                var targetBooks = new HashSet<int>(_paratextService.GetBookList(_userSecret, targetParatextId,
+                    TextType.Target));
                 var sourceBooks = new HashSet<int>(TranslationSuggestionsEnabled
-                    ? (await _paratextService.GetBooksAsync(_userSecret, sourceParatextId))
-                        .Select(bookId => Canon.BookIdToNumber(bookId))
+                    ? _paratextService.GetBookList(_userSecret, targetParatextId, TextType.Source)
                     : Enumerable.Empty<int>());
                 sourceBooks.IntersectWith(targetBooks);
 
-                var targetBooksToDelete = new HashSet<int>(GetBooksToDelete(TextType.Target, targetBooks));
+                var targetBooksToDelete = new HashSet<int>(_projectDoc.Data.Texts.Select(t => t.BookNum)
+                    .Except(targetBooks));
                 var sourceBooksToDelete = new HashSet<int>(TranslationSuggestionsEnabled
-                    ? GetBooksToDelete(TextType.Source, sourceBooks)
+                    ? _projectDoc.Data.Texts.Where(t => t.HasSource).Select(t => t.BookNum).Except(sourceBooks)
                     : Enumerable.Empty<int>());
 
-                _step = 0;
-                _stepCount = (targetBooks.Count * (CheckingEnabled ? 3 : 2)) + (sourceBooks.Count * 2);
+                // delete all data for removed books
                 if (targetBooksToDelete.Count > 0 || sourceBooksToDelete.Count > 0)
                 {
-                    _stepCount += 1;
-
                     // delete source books
                     foreach (int bookNum in sourceBooksToDelete)
                     {
                         TextInfo text = _projectDoc.Data.Texts.First(t => t.BookNum == bookNum);
-                        await DeleteAllTextDataForBookAsync(text, TextType.Source);
+                        await DeleteAllTextDocsForBookAsync(text, TextType.Source);
                     }
                     // delete target books
                     foreach (int bookNum in targetBooksToDelete)
@@ -140,13 +146,12 @@ namespace SIL.XForge.Scripture.Services
                         TextInfo text = _projectDoc.Data.Texts[textIndex];
                         await _projectDoc.SubmitJson0OpAsync(op => op.Remove(pd => pd.Texts, textIndex));
 
-                        await DeleteAllTextDataForBookAsync(text, TextType.Target);
+                        await DeleteAllTextDocsForBookAsync(text, TextType.Target);
                         await DeleteAllQuestionsDocsForBookAsync(text);
                     }
-                    await UpdateProgress();
                 }
 
-                // sync source and target books
+                // update source and target real-time docs
                 foreach (int bookNum in targetBooks)
                 {
                     bool hasSource = sourceBooks.Contains(bookNum);
@@ -157,14 +162,36 @@ namespace SIL.XForge.Scripture.Services
                     else
                         text = _projectDoc.Data.Texts[textIndex];
 
-                    List<Chapter> newChapters = await SyncOrCloneBookUsxAsync(text, TextType.Target, targetParatextId,
-                        false);
+                    // update target text docs
+                    if (!targetTextDocsByBook.TryGetValue(text.BookNum,
+                        out SortedList<int, IDocument<TextData>> targetTextDocs))
+                    {
+                        targetTextDocs = new SortedList<int, IDocument<TextData>>();
+                    }
+
+                    List<Chapter> newChapters = await UpdateTextDocsAsync(text, TextType.Target, targetParatextId,
+                        targetTextDocs);
                     if (hasSource)
                     {
+                        // update source text docs
+                        if (!sourceTextDocsByBook.TryGetValue(text.BookNum,
+                            out SortedList<int, IDocument<TextData>> sourceTextDocs))
+                        {
+                            sourceTextDocs = new SortedList<int, IDocument<TextData>>();
+                        }
                         var chaptersToInclude = new HashSet<int>(newChapters.Select(c => c.Number));
-                        await SyncOrCloneBookUsxAsync(text, TextType.Source, sourceParatextId, true, chaptersToInclude);
+                        await UpdateTextDocsAsync(text, TextType.Source, targetParatextId, sourceTextDocs,
+                            chaptersToInclude);
                     }
-                    await UpdateNotesData(text, newChapters);
+
+                    // update question docs
+                    if (questionDocsByBook.TryGetValue(text.BookNum,
+                        out IReadOnlyList<IDocument<Question>> questionDocs))
+                    {
+                        await UpdateQuestionDocsAsync(questionDocs, newChapters);
+                    }
+
+                    // update project metadata
                     await _projectDoc.SubmitJson0OpAsync(op =>
                     {
                         if (textIndex == -1)
@@ -220,9 +247,6 @@ namespace SIL.XForge.Scripture.Services
             await _notesMapper.InitAsync(_userSecret, _projectSecret, paratextUsers, _projectDoc.Data.ParatextId);
 
             await _projectDoc.SubmitJson0OpAsync(op => op.Set(p => p.Sync.PercentCompleted, 0));
-
-            if (!_fileSystemService.DirectoryExists(WorkingDir))
-                _fileSystemService.CreateDirectory(WorkingDir);
             return true;
         }
 
@@ -231,140 +255,119 @@ namespace SIL.XForge.Scripture.Services
             _conn?.Dispose();
         }
 
-        private async Task<List<Chapter>> SyncOrCloneBookUsxAsync(TextInfo text, TextType textType, string paratextId,
-            bool isReadOnly, ISet<int> chaptersToInclude = null)
+        private void UpdateParatextBook(TextInfo text, string paratextId, SortedList<int, IDocument<TextData>> textDocs)
         {
-            string projectPath = GetProjectPath(textType);
-            if (!_fileSystemService.DirectoryExists(projectPath))
-                _fileSystemService.CreateDirectory(projectPath);
+            string bookText = _paratextService.GetBookText(_userSecret, paratextId, text.BookNum, TextType.Target);
+            var oldUsxDoc = XDocument.Parse(bookText);
+            XDocument newUsxDoc = _deltaUsxMapper.ToUsx(oldUsxDoc, text.Chapters.OrderBy(c => c.Number)
+                .Select(c => new ChapterDelta(c.Number, c.LastVerse, c.IsValid, textDocs[c.Number].Data)));
 
-            string fileName = GetUsxFileName(projectPath, text.BookNum);
-            if (_fileSystemService.FileExists(fileName))
-                return await SyncBookUsxAsync(text, textType, paratextId, fileName, isReadOnly, chaptersToInclude);
-            else
-                return await CloneBookUsxAsync(text, textType, paratextId, fileName, chaptersToInclude);
+            if (!XNode.DeepEquals(oldUsxDoc, newUsxDoc))
+                _paratextService.PutBookText(_userSecret, paratextId, text.BookNum, newUsxDoc.Root.ToString());
         }
 
-        private async Task<List<Chapter>> SyncBookUsxAsync(TextInfo text, TextType textType, string paratextId,
-            string fileName, bool isReadOnly, ISet<int> chaptersToInclude)
+        private async Task UpdateParatextNotesAsync(TextInfo text, IReadOnlyList<IDocument<Question>> questionDocs)
         {
-            SortedList<int, IDocument<Models.TextData>> dbChapterDocs = await FetchTextDocsAsync(text, textType);
+            if (!CheckingEnabled)
+                return;
 
-            string bookId = Canon.BookNumberToId(text.BookNum);
-            string ptBookText = await FetchFromAndUpdateParatextAsync(text,
-                paratextId, fileName, isReadOnly, bookId, dbChapterDocs);
-            await UpdateProgress();
+            // TODO: need to define a data structure for notes instead of XML
+            XElement oldNotesElem;
+            string oldNotesText = _paratextService.GetNotes(_userSecret, _projectDoc.Data.ParatextId, text.BookNum);
+            if (oldNotesText != "")
+                oldNotesElem = ParseText(oldNotesText);
+            else
+                oldNotesElem = new XElement("notes", new XAttribute("version", "1.1"));
 
-            XElement bookTextElem = ParseText(ptBookText);
-            var usxDoc = new XDocument(bookTextElem.Element("usx"));
-            Dictionary<int, ChapterDelta> incomingChapters = _deltaUsxMapper.ToChapterDeltas(usxDoc)
+            XElement notesElem = await _notesMapper.GetNotesChangelistAsync(oldNotesElem, questionDocs);
+
+            if (notesElem.Elements("thread").Any())
+                _paratextService.PutNotes(_userSecret, _projectDoc.Data.ParatextId, notesElem.ToString());
+        }
+
+        private async Task<List<Chapter>> UpdateTextDocsAsync(TextInfo text, TextType textType, string paratextId,
+            SortedList<int, IDocument<TextData>> textDocs, ISet<int> chaptersToInclude = null)
+        {
+            string bookText = _paratextService.GetBookText(_userSecret, paratextId, text.BookNum, textType);
+            var usxDoc = XDocument.Parse(bookText);
+            var tasks = new List<Task>();
+            Dictionary<int, ChapterDelta> deltas = _deltaUsxMapper.ToChapterDeltas(usxDoc)
                 .ToDictionary(cd => cd.Number);
-
-            // Set SF DB to snapshot from Paratext.
-            List<Chapter> chapters = await ChangeDbToNewSnapshotAsync(text,
-                textType, chaptersToInclude, dbChapterDocs, incomingChapters);
-
-            // Save to disk
-            await SaveXmlFileAsync(bookTextElem, fileName);
-
-            await UpdateProgress();
-            return chapters;
-        }
-
-        private async Task<string> FetchFromAndUpdateParatextAsync(TextInfo text, string paratextId,
-            string fileName, bool isReadOnly, string bookId, SortedList<int, IDocument<Models.TextData>> dbChapterDocs)
-        {
-            XElement bookTextElem;
-            string ptBookText;
-            if (isReadOnly)
+            var chapters = new List<Chapter>();
+            foreach (KeyValuePair<int, ChapterDelta> kvp in deltas)
             {
-                ptBookText = await _paratextService.GetBookTextAsync(_userSecret, paratextId, bookId);
-            }
-            else
-            {
-                bookTextElem = await LoadXmlFileAsync(fileName);
-
-                var oldUsxDoc = new XDocument(bookTextElem.Element("usx"));
-                XDocument newUsxDoc = _deltaUsxMapper.ToUsx(oldUsxDoc, text.Chapters.OrderBy(c => c.Number)
-                    .Select(c => new ChapterDelta(c.Number, c.LastVerse, c.IsValid, dbChapterDocs[c.Number].Data)));
-
-                var revision = (string)bookTextElem.Attribute("revision");
-
-                if (XNode.DeepEquals(oldUsxDoc, newUsxDoc))
+                bool addChapter = true;
+                if (textDocs.TryGetValue(kvp.Key, out IDocument<TextData> textDataDoc))
                 {
-                    ptBookText = await _paratextService.GetBookTextAsync(_userSecret, paratextId, bookId);
+                    Delta diffDelta = textDataDoc.Data.Diff(kvp.Value.Delta);
+                    if (diffDelta.Ops.Count > 0)
+                        tasks.Add(textDataDoc.SubmitOpAsync(diffDelta));
+                    textDocs.Remove(kvp.Key);
+                }
+                else if (chaptersToInclude == null || chaptersToInclude.Contains(kvp.Key))
+                {
+                    textDataDoc = GetTextDoc(text, kvp.Key, textType);
+                    async Task createText(int chapterNum, Delta delta)
+                    {
+                        await textDataDoc.FetchAsync();
+                        if (textDataDoc.IsLoaded)
+                            await textDataDoc.DeleteAsync();
+                        await textDataDoc.CreateAsync(new TextData(delta));
+                    }
+                    tasks.Add(createText(kvp.Key, kvp.Value.Delta));
                 }
                 else
                 {
-                    ptBookText = await _paratextService.UpdateBookTextAsync(_userSecret, paratextId, bookId,
-                        revision, newUsxDoc.Root.ToString());
+                    addChapter = false;
                 }
-            }
-            return ptBookText;
-        }
-
-        internal async Task<List<Chapter>> ChangeDbToNewSnapshotAsync(
-            TextInfo text, TextType textType, ISet<int> chaptersToInclude,
-            SortedList<int, IDocument<TextData>> dbChapterDocs, Dictionary<int,
-            ChapterDelta> incomingChapters)
-        {
-            Debug.Assert(dbChapterDocs.All(chapter => chapter.Value.IsLoaded),
-                "Docs must be loaded from the DB.");
-            Debug.Assert(incomingChapters.All(incomingChapter => incomingChapter.Value.Delta != null),
-                "Incoming chapter deltas cannot be null. Maybe DeltaUsxMapper.ToChapterDeltas() has a bug?");
-
-            var tasks = new List<Task>();
-            var chapters = new List<Chapter>();
-            foreach (KeyValuePair<int, ChapterDelta> incomingChapter in incomingChapters)
-            {
-                if (dbChapterDocs.TryGetValue(incomingChapter.Key, out IDocument<Models.TextData> dbChapterDoc))
+                if (addChapter)
                 {
-                    Delta diffDelta = dbChapterDoc.Data.Diff(incomingChapter.Value.Delta);
-                    if (diffDelta.Ops.Count > 0)
+                    chapters.Add(new Chapter
                     {
-                        tasks.Add(dbChapterDoc.SubmitOpAsync(diffDelta));
-                    }
-                    dbChapterDocs.Remove(incomingChapter.Key);
+                        Number = kvp.Key,
+                        LastVerse = kvp.Value.LastVerse,
+                        IsValid = kvp.Value.IsValid
+                    });
                 }
-                else if (chaptersToInclude == null || chaptersToInclude.Contains(incomingChapter.Key))
-                {
-                    // Set database to content from Paratext.
-                    dbChapterDoc = GetTextDoc(text, incomingChapter.Key, textType);
-                    tasks.Add(dbChapterDoc.CreateAsync(new Models.TextData(incomingChapter.Value.Delta)));
-                }
-                chapters.Add(new Chapter
-                {
-                    Number = incomingChapter.Key,
-                    LastVerse = incomingChapter.Value.LastVerse,
-                    IsValid = incomingChapter.Value.IsValid
-                });
             }
-            foreach (KeyValuePair<int, IDocument<Models.TextData>> dbChapterDoc in dbChapterDocs)
-            {
-                tasks.Add(dbChapterDoc.Value.DeleteAsync());
-            }
+            foreach (KeyValuePair<int, IDocument<TextData>> kvp in textDocs)
+                tasks.Add(kvp.Value.DeleteAsync());
             await Task.WhenAll(tasks);
             return chapters;
+        }
+
+        private async Task UpdateQuestionDocsAsync(IReadOnlyList<IDocument<Question>> questionDocs,
+            List<Chapter> newChapters)
+        {
+            // handle deletion of chapters
+            var chapterNums = new HashSet<int>(newChapters.Select(c => c.Number));
+            var tasks = new List<Task>();
+            foreach (IDocument<Question> questionDoc in questionDocs)
+            {
+                if (!chapterNums.Contains(questionDoc.Data.VerseRef.ChapterNum))
+                    tasks.Add(questionDoc.DeleteAsync());
+            }
+            await Task.WhenAll(tasks);
         }
 
         /// <summary>
         /// Fetches all text docs from the database for a book.
         /// </summary>
-        internal async Task<SortedList<int, IDocument<Models.TextData>>> FetchTextDocsAsync(TextInfo text,
+        internal async Task<SortedList<int, IDocument<TextData>>> FetchTextDocsAsync(TextInfo text,
             TextType textType)
         {
-            var textDocs = new SortedList<int, IDocument<Models.TextData>>();
+            var textDocs = new SortedList<int, IDocument<TextData>>();
             var tasks = new List<Task>();
             foreach (Chapter chapter in text.Chapters)
             {
-                IDocument<Models.TextData> textDoc = GetTextDoc(text, chapter.Number, textType);
+                IDocument<TextData> textDoc = GetTextDoc(text, chapter.Number, textType);
                 textDocs[chapter.Number] = textDoc;
                 tasks.Add(textDoc.FetchAsync());
             }
             await Task.WhenAll(tasks);
 
             // Omit items that are not actually in the database.
-            foreach (KeyValuePair<int, IDocument<Models.TextData>> item in textDocs.ToList())
+            foreach (KeyValuePair<int, IDocument<TextData>> item in textDocs.ToList())
             {
                 if (!item.Value.IsLoaded)
                 {
@@ -372,61 +375,6 @@ namespace SIL.XForge.Scripture.Services
                 }
             }
             return textDocs;
-        }
-
-        private async Task<List<Chapter>> CloneBookUsxAsync(TextInfo text, TextType textType, string paratextId,
-            string fileName, ISet<int> chaptersToInclude)
-        {
-            // Remove any stale text_data records that may be in the way.
-            await DeleteAllTextDocsForBookAsync(text, textType);
-
-            string bookText = await _paratextService.GetBookTextAsync(_userSecret, paratextId,
-                Canon.BookNumberToId(text.BookNum));
-            var bookTextElem = ParseText(bookText);
-            await UpdateProgress();
-
-            var usxDoc = new XDocument(bookTextElem.Element("usx"));
-            Dictionary<int, ChapterDelta> deltas = _deltaUsxMapper.ToChapterDeltas(usxDoc)
-                .ToDictionary(cd => cd.Number);
-            var tasks = new List<Task>();
-            var chapters = new List<Chapter>();
-            foreach (KeyValuePair<int, ChapterDelta> kvp in deltas)
-            {
-                if (chaptersToInclude != null && !chaptersToInclude.Contains(kvp.Key))
-                    continue;
-
-                async Task createText(int chapterNum, Delta delta)
-                {
-                    IDocument<Models.TextData> textDataDoc = GetTextDoc(text, chapterNum, textType);
-                    await textDataDoc.FetchAsync();
-                    if (textDataDoc.IsLoaded)
-                        await textDataDoc.DeleteAsync();
-                    await textDataDoc.CreateAsync(new Models.TextData(delta));
-                }
-                tasks.Add(createText(kvp.Key, kvp.Value.Delta));
-                chapters.Add(new Chapter
-                {
-                    Number = kvp.Key,
-                    LastVerse = kvp.Value.LastVerse,
-                    IsValid = kvp.Value.IsValid
-                });
-            }
-            await Task.WhenAll(tasks);
-
-            await SaveXmlFileAsync(bookTextElem, fileName);
-
-            await UpdateProgress();
-            return chapters;
-        }
-
-        /// <summary>
-        /// Deletes the cached USX file from the disk and all text docs from the database for a book.
-        /// </summary>
-        private async Task DeleteAllTextDataForBookAsync(TextInfo text, TextType textType)
-        {
-            string projectPath = GetProjectPath(textType);
-            _fileSystemService.DeleteFile(GetUsxFileName(projectPath, text.BookNum));
-            await DeleteAllTextDocsForBookAsync(text, textType);
         }
 
         /// <summary>
@@ -438,42 +386,6 @@ namespace SIL.XForge.Scripture.Services
             foreach (Chapter chapter in text.Chapters)
                 tasks.Add(DeleteTextDocAsync(text, chapter.Number, textType));
             await Task.WhenAll(tasks);
-        }
-
-        private async Task UpdateNotesData(TextInfo text, List<Chapter> newChapters)
-        {
-            IReadOnlyList<IDocument<Question>> allQuestionDocs = await FetchQuestionDocsAsync(text);
-
-            // handle deletion of chapters
-            var chapterNums = new HashSet<int>(newChapters.Select(c => c.Number));
-            var tasks = new List<Task>();
-            foreach (IDocument<Question> questionDoc in allQuestionDocs)
-            {
-                if (!chapterNums.Contains(questionDoc.Data.VerseRef.ChapterNum))
-                    tasks.Add(questionDoc.DeleteAsync());
-            }
-            await Task.WhenAll(tasks);
-
-            if (CheckingEnabled)
-            {
-                XElement oldNotesElem;
-                string oldNotesText = await _paratextService.GetNotesAsync(_userSecret, _projectDoc.Data.ParatextId,
-                    Canon.BookNumberToId(text.BookNum));
-                if (oldNotesText != "")
-                    oldNotesElem = ParseText(oldNotesText);
-                else
-                    oldNotesElem = new XElement("notes", new XAttribute("version", "1.1"));
-
-                XElement notesElem = await _notesMapper.GetNotesChangelistAsync(oldNotesElem, allQuestionDocs);
-
-                if (notesElem.Elements("thread").Any())
-                {
-                    await _paratextService.UpdateNotesAsync(_userSecret, _projectDoc.Data.ParatextId,
-                        notesElem.ToString());
-                }
-
-                await UpdateProgress();
-            }
         }
 
         private async Task<IReadOnlyList<IDocument<Question>>> FetchQuestionDocsAsync(TextInfo text)
@@ -529,20 +441,6 @@ namespace SIL.XForge.Scripture.Services
             await Task.WhenAll(tasks);
         }
 
-        /// <summary>
-        /// Gets all books that need to be deleted.
-        /// </summary>
-        private IEnumerable<int> GetBooksToDelete(TextType textType, IEnumerable<int> existingBooks)
-        {
-            string projectPath = GetProjectPath(textType);
-            var booksToDelete = new HashSet<int>(_fileSystemService.DirectoryExists(projectPath)
-                ? _fileSystemService.EnumerateFiles(projectPath)
-                    .Select(p => Canon.BookIdToNumber(Path.GetFileNameWithoutExtension(p)))
-                : Enumerable.Empty<int>());
-            booksToDelete.ExceptWith(existingBooks);
-            return booksToDelete;
-        }
-
         private async Task CompleteSync(bool successful)
         {
             if (_projectDoc == null || _projectSecret == null)
@@ -588,64 +486,32 @@ namespace SIL.XForge.Scripture.Services
             }
         }
 
-        private string GetProjectPath(TextType textType)
+        private IDocument<TextData> GetTextDoc(TextInfo text, int chapter, TextType textType)
         {
-            string textTypeDir;
-            switch (textType)
-            {
-                case TextType.Source:
-                    textTypeDir = "source";
-                    break;
-                case TextType.Target:
-                    textTypeDir = "target";
-                    break;
-                default:
-                    throw new InvalidEnumArgumentException(nameof(textType), (int)textType, typeof(TextType));
-            }
-            return Path.Combine(WorkingDir, _projectDoc.Id, textTypeDir);
-        }
-
-        private static string GetUsxFileName(string projectPath, int bookNum)
-        {
-            return Path.Combine(projectPath, Canon.BookNumberToId(bookNum) + ".xml");
-        }
-
-        private async Task<XElement> LoadXmlFileAsync(string fileName)
-        {
-            using (Stream stream = _fileSystemService.OpenFile(fileName, FileMode.Open))
-            {
-                return await XElement.LoadAsync(stream, LoadOptions.PreserveWhitespace, CancellationToken.None);
-            }
-        }
-
-        private async Task SaveXmlFileAsync(XElement bookTextElem, string fileName)
-        {
-            using (Stream stream = _fileSystemService.CreateFile(fileName))
-            {
-                await bookTextElem.SaveAsync(stream, SaveOptions.DisableFormatting, CancellationToken.None);
-            }
-        }
-
-        private IDocument<Models.TextData> GetTextDoc(TextInfo text, int chapter, TextType textType)
-        {
-            return _conn.Get<Models.TextData>(TextData.GetTextDocId(_projectDoc.Id, text.BookNum, chapter, textType));
+            return _conn.Get<TextData>(TextData.GetTextDocId(_projectDoc.Id, text.BookNum, chapter, textType));
         }
 
         private async Task DeleteTextDocAsync(TextInfo text, int chapter, TextType textType)
         {
-            IDocument<Models.TextData> textDoc = GetTextDoc(text, chapter, textType);
+            IDocument<TextData> textDoc = GetTextDoc(text, chapter, textType);
             await textDoc.FetchAsync();
             if (textDoc.IsLoaded)
                 await textDoc.DeleteAsync();
         }
 
-        private async Task UpdateProgress()
+        private SyncProgress UseNewProgress()
         {
-            if (_projectDoc == null)
-                return;
-            _step++;
-            double percentCompleted = Math.Round((double)_step / _stepCount, 2, MidpointRounding.AwayFromZero);
-            await _projectDoc.SubmitJson0OpAsync(op => op.Set(pd => pd.Sync.PercentCompleted, percentCompleted));
+            var progress = new SyncProgress();
+            progress.ProgressUpdated += (object sender, EventArgs e) =>
+            {
+                if (_projectDoc == null)
+                    return;
+                double percentCompleted = progress.ProgressValue;
+                if (percentCompleted >= 0)
+                    _projectDoc.SubmitJson0OpAsync(op => op.Set(pd => pd.Sync.PercentCompleted, percentCompleted));
+            };
+
+            return progress;
         }
 
         private class ChapterEqualityComparer : IEqualityComparer<Chapter>
